requirements.txt
Appium-Python-Client==0.49
selenium==4.0.0a3